# api/swagger/swagger.yaml
swagger: "2.0"
info:
  title: Vitality
  version: v1
#host: my.remote.host
host: localhost:3000
basePath: /
schemes:
  - http
  - https
consumes:
- application/json
- application/javascript
produces:
- application/json
- application/javascript
- text/javascript
paths:
...
  /getcommits/{id}:
    x-swagger-router-controller: core_getcommits
    get:
      summary: Get commit counts for a specific repository
      description: |
        Returns an array of commit counts, by month, for a specific repository.  The most recent six months of data is returned by default, up to 12 months can be returned via the 'count' parameter.
      operationId: getcommits
      parameters:
        - name: id
          in: path
          description: The GitHub ID of the repository to be queried
          required: true
          type: integer
        - name: callback
          in: query
          description: Passed by .ajax call for cross-domain requests
          required: false
          type: string
        - name: count
          in: query
          description: An integer between 1 and 12, inclusive.
          required: false
          type: integer
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/getcommitsResponse"
        400:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  ErrorResponse:
    type: string
    required:
      - success
      - message
    properties:
      success:
        type: string
      message:
        type: string
      nativemessage:
        type: string
      nativecode:
        type: integer
...
  CommitData:
    type: object
    required:
      - month
      - count
    properties:
      month:
        type: string
      count:
        type: integer
  getcommitsResponse:
    #type: object  // replaced with type:string to allow JSONP response
    type: string
    required:
      - success
      - data
    properties:
      success:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/CommitData"
...